% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorhere.R
\name{anchorhere}
\alias{anchorhere}
\title{Drop the anchor to specific directory}
\usage{
anchorhere(proj.nm = date(), anchor = "_proj_version_", path = "./")
}
\arguments{
\item{proj.nm}{character First line in anchor file.}

\item{anchor}{character Anchor filename.}

\item{path}{Path. Directory to drop anchor.}
}
\description{
Drop the anchor to specific directory
}
\examples{
\dontrun{anchorhere("my project!")}
}
