% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ishere.R
\name{ishere}
\alias{ishere}
\title{Get directory}
\usage{
ishere(proj.nm = "i am here", anchor = "_proj_version_", stop.here = "/mnt")
}
\arguments{
\item{proj.nm}{character First line in anchor file.}

\item{anchor}{character Anchor filename.}

\item{stop.here}{Path. Top path to stop search.}
}
\value{
character Path.
}
\description{
Get directory
}
\examples{
\dontrun{ishere("my project.")}
}
