% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tothere.R
\name{tothere}
\alias{tothere}
\title{Return target subpath of project root of same depth with current directory}
\usage{
tothere(anchor.path, to.branch = "data", pwd.branch = "script")
}
\arguments{
\item{anchor.path}{character Project root path.}

\item{to.branch}{character Target subpath of project root path.}

\item{pwd.branch}{character Current subpath of project root path.}
}
\value{
character Path
}
\description{
Return target subpath of project root of same depth with current directory
}
\examples{
\dontrun{tothere(ishere("my project!"))}
}
