% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whereis.R
\name{whereis}
\alias{whereis}
\title{Find files under target path}
\usage{
whereis(path = "./", pattern, ..., absolute.nm = TRUE, ignore.case = FALSE)
}
\arguments{
\item{path}{character Path.}

\item{pattern}{character File pattern}

\item{...}{character Specify subpathes to search.}

\item{absolute.nm}{logical 'pattern' is full name or not.}

\item{ignore.case}{logical Should pattern-matching be case-insensitive?}
}
\value{
character Path.
}
\description{
Find files under target path
}
\examples{
\dontrun{whereis(pattern = "_proj_version_")}
\dontrun{whereis(pattern = "_proj_version_", "here", "there")}
}
